\name{NetworkAnalysis-internal}
\alias{vec2sym2}
\alias{belong}
\alias{transposing}
\alias{cost2adj}
\alias{fdr}
\alias{ge.ind}
\alias{le.ind}
\alias{naObj}
\alias{top.metric}
\alias{vector.thres}

\title{Internal NetworkAnalysis Functions}
\description{Internal NetworkAnalysis functions}
\details{
These are not to be called by the user (or in some cases are just
waiting for proper documentation to be written :).}
\keyword{internal}